Misc. fixes, also expanded out PIUIO to template extra sensors
git-svn-id: https://openitg.svn.sourceforge.net/svnroot/openitg@87 83fadc84-e282-4d84-a09a-c4228d6ae7e5

+++ src/arch/InputHandler/InputHandler_Linux_PIUIO.h
@@ -25,8 +25,8 @@ private:
 	bool m_bFoundDevice;
 	bool m_bShutdown;
 
-	uint32_t m_iInputData;
-	uint32_t m_iLastInputData;
+	uint64_t m_iInputData;
+	uint64_t m_iLastInputData;
 
 	uint32_t m_iLightData;
 	uint32_t m_iLastLightData;
